Break up changelog package
The changelog implementation is growing. This commit breaks out the
builder in to it's own file. This is purely for organizational
purposes and has no impact on the functionality of the changelog
builder.
